This code is correct
The loop takes long time to run for bigger number. Instead of writing a loop the every number in the pile[] is substracted by 2 and the result is added to turn variable. The summation of the result in turn variable is then proceeded further to check if CHEF wins or CHEFINA wins